refactor: split ODK Collect views
